% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add.blank.lines}
\alias{add.blank.lines}
\title{Add trailing and leading blank line}
\usage{
add.blank.lines(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Adds a line break before \emph{and} after the character string(s).
}
