\name{add.lattice.xsubticks}
\alias{add.lattice.xsubticks}
\alias{add.lattice.ysubticks}
\title{Add subticks}
\usage{
  add.lattice.xsubticks(lim, ..., n = 5, n2 = n * 5,
    min.n2 = n + 5)
}
\arguments{
  \item{lim}{see
  \code{latticeExtra::xscale.components.subticks}}

  \item{...}{see
  \code{latticeExtra::xscale.components.subticks}}

  \item{n}{see
  \code{latticeExtra::xscale.components.subticks}}

  \item{n2}{see
  \code{latticeExtra::xscale.components.subticks}}

  \item{min.n2}{see
  \code{latticeExtra::xscale.components.subticks}}
}
\description{
  See \code{latticeExtra::xscale.components.subticks}.
}
\references{
  This is copied from
  \code{latticeExtra::xscale.components.subticks}.
}
\keyword{internal}

