% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add.minor.ticks}
\alias{add.minor.ticks}
\title{Minor tick marks}
\usage{
add.minor.ticks(nx = 4, ny = 4, grid = TRUE)
}
\arguments{
\item{nx}{intervals to split on x axis}

\item{ny}{intervals to split on x axis}

\item{grid}{show grid lines at minor ticks?}
}
\description{
Adds minor tick marks to an existing \code{base} plot.
}
\references{
This function is a forked version of \code{Hmisc::minor.tick}
}
\keyword{internal}
