% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.significance.stars}
\alias{add.significance.stars}
\title{Add significance stars}
\usage{
add.significance.stars(p)
}
\arguments{
\item{p}{numeric vector or tabular data}
}
\value{
character vector
}
\description{
This function adds significance stars to passed \code{p} value(s) as: one star for value below \code{0.05}, two for \code{0.01} and three for \code{0.001}.
}

