\name{cache.off}
\alias{cache.off}
\alias{cache.on}
\title{Toggle cache}
\usage{
  cache.on()

  cache.off()
}
\description{
  This function is just a wrapper around
  \code{\link{evalsOptions}} to switch pander's cache on or
  off easily, which might be handy in some brew documents
  to prevent repetitive strain injury :)
}

