% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_caption}
\alias{check_caption}
\title{Check if caption is valid}
\usage{
check_caption(caption)
}
\arguments{
\item{caption}{R object to check}
}
\value{
boolean
}
\description{
Check if caption is valid
}
\keyword{internal}
