% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_digits}
\alias{check_digits}
\title{Check if vector parameter for round/digits and adjust accodingly}
\usage{
check_digits(param, name, n)
}
\arguments{
\item{param}{vector to be checked}

\item{name}{parameter name}

\item{n}{needed size of vector}
}
\value{
original vector if size is good, vector of default values otherwise
}
\description{
Check if vector parameter for round/digits and adjust accodingly
}
\keyword{internal}
