% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coef_mat}
\alias{coef_mat}
\title{Calculate coef matrix for models from rms package
Forked from prModFit from rms}
\usage{
coef_mat(obj, coefs)
}
\arguments{
\item{obj}{object list}

\item{coefs}{numeric value if to print only the first n regression coefficients in the model.}
}
\value{
coeficients matrix
}
\description{
Calculate coef matrix for models from rms package
Forked from prModFit from rms
}
