\name{emphasize.rows}
\alias{emphasize.cells}
\alias{emphasize.cols}
\alias{emphasize.rows}
\alias{emphasize.strong.cells}
\alias{emphasize.strong.cols}
\alias{emphasize.strong.rows}
\title{Emphasize rows/columns/cells}
\usage{
  emphasize.rows(x)

  emphasize.cols(x)

  emphasize.cells(x)

  emphasize.strong.rows(x)

  emphasize.strong.cols(x)

  emphasize.strong.cells(x)
}
\arguments{
  \item{x}{vector of row/columns indexes or an array like
  returned by \code{which(..., arr.ind = TRUE)}}
}
\description{
  TODO
}

