\name{get.emphasize}
\alias{get.emphasize}
\title{Get emphasize params from internal buffer}
\usage{
  get.emphasize(df)
}
\arguments{
  \item{df}{tabular data}
}
\value{
  R object passed as \code{df} with possibly added
  \code{attr}s captured from internal buffer
}
\description{
  And truncate content.
}
\keyword{internal}

