% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has.rownames}
\alias{has.rownames}
\title{Check if rownames are available}
\usage{
has.rownames(x)
}
\arguments{
\item{x}{a tabular-like R object}
}
\value{
\code{TRUE} OR \code{FALSE}
}
\description{
Dummy helper to check if the R object has real rownames or not.
}

