% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multitable}
\alias{multitable}
\title{Create a multitable used for rendering objects from rms package}
\usage{
multitable(v)
}
\arguments{
\item{v}{list of vectors/lists to be merge in the table}
}
\value{
data.frame in specified format
}
\description{
When ols/lrm/orm from rms package get rendered, main statistics are group in table of tables.
Since pandoc doesn't support row or col-span,
we chose to group those statistics in a column each.
This function takes care of that
}
\examples{
pander:::multitable(list(list(a=1, b=2),list(c=3, d=4)))
}
\keyword{internal}
