\name{open.file.in.OS}
\alias{open.file.in.OS}
\title{Open file}
\usage{
  open.file.in.OS(f)
}
\arguments{
  \item{f}{file (with full path)}
}
\description{
  Tries to open a file with operating system's default
  program.
}
\references{
  This function is a fork of David Hajage's \code{convert}
  function:
  \url{https://github.com/eusebe/ascii/blob/master/R/export.r}
}

