\name{p}
\alias{p}
\title{Pretty print vectors}
\usage{
  p(x)
}
\arguments{
  \item{x}{vector}
}
\value{
  character string
}
\description{
  Pretty print vectors
}
\examples{
p(1:5)
}
\references{
  Credit goes to Aleksandar Blagotić for the idea and
  implementation in rapport package.
  \url{http://github.com/aL3xa/rapport/blob/master/R/rp_helpers.R}
}

