% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot.R
\name{palette_pander}
\alias{palette_pander}
\title{Color palette from the pander package}
\usage{
palette_pander(n, random_order = FALSE,
  cols = panderOptions("graph.colors"))
}
\arguments{
\item{n}{number of colors}

\item{random_order}{if the palette should be reordered randomly before
rendering each plot to get colorful images}

\item{cols}{vector of colors to use}
}
\description{
The \pkg{pander} ships with a default colorblind and printer-friendly
color palette borrowed from \url{http://jfly.iam.u-tokyo.ac.jp/color/}.
}
\examples{
palette_pander(TRUE)
}

