% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.aovlist}
\alias{pander.aovlist}
\title{Pander method for aovlist class}
\usage{
\method{pander}{aovlist}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an aovlist object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an aovlist object in Pandoc's markdown.
}
