% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.Arima}
\alias{pander.Arima}
\title{Prints an arima object from stats package in Pandoc's markdown.}
\usage{
\method{pander}{Arima}(x, digits = panderOptions("digits"), se = TRUE,
  ...)
}
\arguments{
\item{x}{an arima object}

\item{digits}{number of digits of precision}

\item{se}{if to include standard error in coefficients table (default \code{TRUE})}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an arima object from stats package in Pandoc's markdown.
}
