% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.call}
\alias{pander.call}
\title{Pander method for call class}
\usage{
\method{pander}{call}(x, ...)
}
\arguments{
\item{x}{a call object}

\item{...}{optional parameters passed to raw \code{pandoc.formula} function}
}
\description{
Prints a call object in Pandoc's markdown.
}
