% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.cast_df}
\alias{pander.cast_df}
\title{Pander method for cast_df class}
\usage{
\method{pander}{cast_df}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a cast_df object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a cast_df object in Pandoc's markdown.
}
