% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.character}
\alias{pander.character}
\title{Pander method for character class}
\usage{
\method{pander}{character}(x, ...)
}
\arguments{
\item{x}{a character object}

\item{...}{igroned parameters}
}
\description{
Prints a character class in Pandoc's markdown.
}

