% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.clogit}
\alias{pander.clogit}
\title{Pander method for clogit class}
\usage{
\method{pander}{clogit}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an clogit object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a clogit object in Pandoc's markdown.
}
