% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.cph}
\alias{pander.cph}
\title{Prints an cph object from rms package in Pandoc's markdown.}
\usage{
\method{pander}{cph}(x, table = TRUE, conf.int = FALSE, coefs = TRUE,
  ...)
}
\arguments{
\item{x}{an cph object}

\item{table}{if to print event frequency statistics. default(\code{TRUE})}

\item{conf.int}{set to e.g. .95 to print 0.95 confidence intervals on simple hazard ratios (which are usually meaningless as one-unit changes are seldom relevant and most models contain multiple terms per predictor)}

\item{coefs}{if to the table of model coefficients, standard errors, etc. default(\code{TRUE})}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an cph object from rms package in Pandoc's markdown.
}
