% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.data.frame}
\alias{pander.data.frame}
\title{Pander method for data.frame class}
\usage{
\method{pander}{data.frame}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a data.frame object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a data.frame object in Pandoc's markdown.
}
