% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.data.table}
\alias{pander.data.table}
\title{Pander method for data.table class}
\usage{
\method{pander}{data.table}(x, caption = attr(x, "caption"),
  keys.as.row.names = TRUE, ...)
}
\arguments{
\item{x}{a data.table object}

\item{caption}{caption (string) to be shown under the table}

\item{keys.as.row.names}{controls whether to use data.table key as row names when calling pandoc.table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a data.table object in Pandoc's markdown. Data.tables drop attributes (like row names) when called.
}
