% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.default}
\alias{pander.default}
\title{Default Pander method}
\usage{
\method{pander}{default}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{optional parameters passed to raw \code{pandoc.list} function}
}
\description{
Method to be used, when no exact S3 method for given object is found. Tries to render object as a list
}
