% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.describe}
\alias{pander.describe}
\title{Pander method for describe class}
\usage{
\method{pander}{describe}(x, caption = attr(x, "caption"), short = TRUE,
  split.tables = 60, ...)
}
\arguments{
\item{x}{an describe object}

\item{caption}{caption (string) to be shown under the table}

\item{short}{(default:\code{TRUE}) if to use consise output}

\item{split.tables}{(default:\code{60}) split.tables param for pandoc.table function}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a describe object in Pandoc's markdown.
}

