% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.ets}
\alias{pander.ets}
\title{Prints an ets object from forecast package in Pandoc's markdown.}
\usage{
\method{pander}{ets}(x, digits = panderOptions("digits"), ...)
}
\arguments{
\item{x}{an ets object}

\item{digits}{number of digits of precision}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an ets object from forecast package in Pandoc's markdown.
}
