% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.factor}
\alias{pander.factor}
\title{Pander method for factor class}
\usage{
\method{pander}{factor}(x, ...)
}
\arguments{
\item{x}{a factor object}

\item{...}{igroned parameters}
}
\description{
Prints a factor object in Pandoc's markdown.
}
