% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.formula}
\alias{pander.formula}
\title{Pander method for formula class}
\usage{
\method{pander}{formula}(x, max.width = 80, caption = attr(x,
  "caption"), ...)
}
\arguments{
\item{x}{a formula object}

\item{max.width}{maximum width in characters per line}

\item{caption}{caption (string) to be shown under the formula}

\item{...}{optional parameters passed to raw \code{pandoc.formula} function}
}
\description{
Prints a formula object in Pandoc's markdown.
}
