% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.ftable}
\alias{pander.ftable}
\title{Pander method for ftable class}
\usage{
\method{pander}{ftable}(x, ...)
}
\arguments{
\item{x}{a ftable object}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a ftable object in Pandoc's markdown.
}

