% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.function}
\alias{pander.function}
\title{Pander method for function class}
\usage{
\method{pander}{function}(x, add.name = FALSE, verbatim = TRUE,
  syntax.highlighting = FALSE, ...)
}
\arguments{
\item{x}{an function object}

\item{add.name}{(defaut:\code{FALSE}) if to add function name to output or just to print a body}

\item{verbatim}{(defaut:\code{TRUE}) if to add tabulation, so pandoc conversion will rander it properly}

\item{syntax.highlighting}{(defaut:\code{FALSE}) if to add hyghlighting tag for R syntax}

\item{...}{ignored parameters}
}
\description{
Prints an function object in Pandoc's markdown.
}
