% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.glm}
\alias{pander.glm}
\title{Pander method for summary.glm class}
\usage{
\method{pander}{glm}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a summary.glm object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.glm object in Pandoc's markdown.
}
