% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.gtable}
\alias{pander.gtable}
\title{Pander method for gtable class}
\usage{
\method{pander}{gtable}(x, zsort = FALSE, ...)
}
\arguments{
\item{x}{an gtable object}

\item{zsort}{Sort by z values? Default \code{FALSE}}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Renders an gtable object in Pandoc's markdown.
}
