% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.htest}
\alias{pander.htest}
\title{Pander method for htest class}
\usage{
\method{pander}{htest}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a htest object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a htest object in Pandoc's markdown.
}

