% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.irts}
\alias{pander.irts}
\title{Prints an irts object from tseries package in Pandoc's markdown.}
\usage{
\method{pander}{irts}(x, caption = attr(x, "caption"),
  format = panderOptions("date"), ...)
}
\arguments{
\item{x}{an irts object}

\item{caption}{caption (string) to be shown under the table}

\item{format}{string passed to format when printing dates (POSIXct or POSIXt)}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an irts object from tseries package in Pandoc's markdown.
}
