% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.list}
\alias{pander.list}
\title{Pander method for list class}
\usage{
\method{pander}{list}(x, ...)
}
\arguments{
\item{x}{a list object}

\item{...}{ignored parameters}
}
\description{
Prints a list object in Pandoc's markdown.
}
