% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.lm}
\alias{pander.lm}
\title{Pander method for summary.glm class}
\usage{
\method{pander}{lm}(x, caption = attr(x, "caption"), covariate.labels, omit,
  ...)
}
\arguments{
\item{x}{a summary.glm object}

\item{caption}{caption (string) to be shown under the table}

\item{covariate.labels}{vector to replace covariate lables in the table}

\item{omit}{vector of variable to omit for priting in resulting table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.glm object in Pandoc's markdown.
}

