% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.logical}
\alias{pander.logical}
\title{Pander method for logical class}
\usage{
\method{pander}{logical}(x, ...)
}
\arguments{
\item{x}{a logical object}

\item{...}{ignored parameters}
}
\description{
Prints a logical object in Pandoc's markdown.
}
