% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.manova}
\alias{pander.manova}
\title{Pander method for manova class}
\usage{
\method{pander}{manova}(x, caption = attr(x, "caption"),
  add.significance.stars = FALSE, ...)
}
\arguments{
\item{x}{an manovv object}

\item{caption}{caption (string) to be shown under the table}

\item{add.significance.stars}{if significance stars should be shown for P value}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an manova object in Pandoc's markdown.
}
