% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.matrix}
\alias{pander.matrix}
\title{Pander method for matrix class}
\usage{
\method{pander}{matrix}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a matrix object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a matrix object in Pandoc's markdown.
}

