% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.microbenchmark}
\alias{pander.microbenchmark}
\title{Pander method for microbenchmark class}
\usage{
\method{pander}{microbenchmark}(x, caption = attr(x, "caption"),
  expr.labels, unit, ...)
}
\arguments{
\item{x}{an microbenchmark object}

\item{caption}{caption (string) to be shown under the table}

\item{expr.labels}{expression labels that will replace default ones (similar to rownames, which microbenchmark class table does not have)}

\item{unit}{units in which values should be printed (for example second, microseconds, etc.). Should be one of ns, us, ms, s, t, hz, khz, mhz, eps, f}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an microbenchmark object in Pandoc's markdown.
}
