% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.mtable}
\alias{pander.mtable}
\title{Pander method for mtable class}
\usage{
\method{pander}{mtable}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a mtable object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a mtable object in Pandoc's markdown.
}

