% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.name}
\alias{pander.name}
\title{Pander method for name class}
\usage{
\method{pander}{name}(x, ...)
}
\arguments{
\item{x}{a name language object}
}
\description{
Prints a call object in Pandoc's markdown.
}

