% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.nls}
\alias{pander.nls}
\title{Prints an nls object from stats package in Pandoc's markdown.}
\usage{
\method{pander}{nls}(x, digits = panderOptions("digits"),
  show.convergence = FALSE, ...)
}
\arguments{
\item{x}{an nls object}

\item{digits}{number of digits of precision}

\item{show.convergence}{(defaut:\code{FALSE}) if to print convergence info}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an nls object from stats package in Pandoc's markdown.
}
