% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.NULL}
\alias{pander.NULL}
\title{Pander method for a NULL object}
\usage{
\method{pander}{NULL}(x, ...)
}
\arguments{
\item{x}{a NULL object}

\item{...}{ignored parameters}
}
\description{
Prints a NULL object in Pandoc's markdown.
}

