% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.numeric}
\alias{pander.numeric}
\title{Pander method for numeric class}
\usage{
\method{pander}{numeric}(x, ...)
}
\arguments{
\item{x}{a numeric object}

\item{...}{igroned parameter}
}
\description{
Prints a numeric class in Pandoc's markdown.
}

