% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.ols}
\alias{pander.ols}
\title{Prints an ols object from rms package in Pandoc's markdown.}
\usage{
\method{pander}{ols}(x, long = FALSE, coefs = TRUE,
  digits = panderOptions("digits"), round = panderOptions("round"),
  ...)
}
\arguments{
\item{x}{an ols object}

\item{long}{if to print the correlation matrix of parameter estimates. default(\code{FALSE})}

\item{coefs}{if to the table of model coefficients, standard errors, etc. default(\code{TRUE})}

\item{digits}{passed to \code{format}. Can be a vector specifying values for each column (has to be the same length as number of columns).}

\item{round}{passed to \code{round}. Can be a vector specifying values for each column (has to be the same length as number of columns). Values for non-numeric columns will be disregarded.}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an ols object from rms package in Pandoc's markdown.
}
