% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.orm}
\alias{pander.orm}
\title{Prints an orm object from rms package in Pandoc's markdown.}
\usage{
\method{pander}{orm}(x, coefs = TRUE, intercepts = x$non.slopes < 10,
  ...)
}
\arguments{
\item{x}{an orm object}

\item{coefs}{if to the table of model coefficients, standard errors, etc. default(\code{TRUE})}

\item{intercepts}{if to print intercepts, by default, intercepts are only printed if there are fewer than 10 of them}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an orm object from rms package in Pandoc's markdown.
}
