% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.polr}
\alias{pander.polr}
\title{Prints an polr object from MASS package in Pandoc's markdown.}
\usage{
\method{pander}{polr}(x, ...)
}
\arguments{
\item{x}{an polr object}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an polr object from MASS package in Pandoc's markdown.
}
