% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.POSIXlt}
\alias{pander.POSIXlt}
\title{Pander method for POSIXlt class}
\usage{
\method{pander}{POSIXlt}(x, ...)
}
\arguments{
\item{x}{a POSIXlt object}

\item{...}{optional parameters passed to raw \code{pandoc.date} function}
}
\description{
Prints a POSIXlt object in Pandoc's markdown.
}

