% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.prcomp}
\alias{pander.prcomp}
\title{Pander method for prcomp class}
\usage{
\method{pander}{prcomp}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a prcomp object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a prcomp object in Pandoc's markdown.
}

