% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.randomForest}
\alias{pander.randomForest}
\title{Pander method for randomForest class}
\usage{
\method{pander}{randomForest}(x, digits = panderOptions("digits"), ...)
}
\arguments{
\item{x}{an randomForest object}

\item{digits}{number of digits of precision}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Renders an randomForest object in Pandoc's markdown.
}
