% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.rapport}
\alias{pander.rapport}
\title{Pander method for rapport class}
\usage{
\method{pander}{rapport}(x, ...)
}
\arguments{
\item{x}{a rapport object}

\item{...}{ignored parameters}
}
\description{
Prints a rapport object in Pandoc's markdown.
}

