% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.return}
\alias{pander.return}
\title{Deprecated pander.return}
\usage{
pander.return(...)
}
\arguments{
\item{...}{everything passed to \code{pander}}
}
\description{
This is a wrapper function around \code{pander} but instead of printing to \code{stdout}, this function returns a character vector of the captured lines.
}
\seealso{
pander
}

