% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.sessionInfo}
\alias{pander.sessionInfo}
\title{Pander method for sessionInfo class}
\usage{
\method{pander}{sessionInfo}(x, locale = TRUE, compact = TRUE, ...)
}
\arguments{
\item{x}{an sessionInfo object}

\item{locale}{(defaut:\code{TRUE}) if to print locale output}

\item{compact}{(defaut:\code{TRUE}) if output shoud be compact (ommiting extra line breaks and spaces, inline printing of lists)}

\item{...}{ignored parameters}
}
\description{
Prints an sessionInfo object in Pandoc's markdown.
}

