% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.smooth.spline}
\alias{pander.smooth.spline}
\title{Pander method for smooth.spline class}
\usage{
\method{pander}{smooth.spline}(x, ...)
}
\arguments{
\item{x}{an smooth.spline object}

\item{...}{igroned parameters}
}
\description{
Prints an smooth.spline object in Pandoc's markdown.
}

