% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.stat.table}
\alias{pander.stat.table}
\title{Pander method for stat.table class}
\usage{
\method{pander}{stat.table}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an stat.table object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an stat.table object in Pandoc's markdown.
}
