% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.aov}
\alias{pander.summary.aov}
\title{Pander method for summary.aov class}
\usage{
\method{pander}{summary.aov}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a summary.aov object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.aov object in Pandoc's markdown.
}
