% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.summary.aovlist}
\alias{pander.summary.aovlist}
\title{Pander method for summary.aovlist class}
\usage{
\method{pander}{summary.aovlist}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a summary.aovlist object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.aovlist object in Pandoc's markdown.
}

