% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.glm}
\alias{pander.summary.glm}
\title{Pander method for summary.glm class}
\usage{
\method{pander}{summary.glm}(x, caption = attr(x, "caption"),
  covariate.labels, omit, summary = TRUE, ...)
}
\arguments{
\item{x}{an summary.glm object}

\item{caption}{caption (string) to be shown under the table}

\item{covariate.labels}{vector to replace covariate lables in the table}

\item{omit}{vector of variable to omit for priting in resulting table}

\item{summary}{(defaut:\code{TRUE}) if used for summary.lm or lm}

\item{...}{optional parameters passed to special methods and/or raw \code{pandoc.*} functions}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Prints a summary.glm object in Pandoc's markdown.
}
