% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.lm}
\alias{pander.summary.lm}
\title{Pander method for summary.lm class}
\usage{
\method{pander}{summary.lm}(x, caption = attr(x, "caption"),
  covariate.labels, omit, summary = TRUE,
  add.significance.stars = FALSE, move.intercept = FALSE, ...)
}
\arguments{
\item{x}{an summary.lm object}

\item{caption}{caption (string) to be shown under the table}

\item{covariate.labels}{vector to replace covariate lables in the table}

\item{omit}{vector of variable to omit for priting in resulting table}

\item{summary}{(defaut:\code{TRUE}) if used for summary.lm or lm}

\item{add.significance.stars}{if significance stars should be shown for P value}

\item{move.intercept}{by default, the Intercept is the first coefficient in the table, which can be moved to the bottom of the table}

\item{...}{optional parameters passed to special methods and/or raw \code{pandoc.*} functions}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Prints a summary.lm object in Pandoc's markdown.
}
