% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.manova}
\alias{pander.summary.manova}
\title{Prints an summary.manova object from stats package in Pandoc's markdown.}
\usage{
\method{pander}{summary.manova}(x, caption = attr(x, "caption"),
  add.significance.stars = FALSE, ...)
}
\arguments{
\item{x}{an summary.manova object}

\item{caption}{caption (string) to be shown under the table}

\item{add.significance.stars}{if significance stars should be shown for P value}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an summary.manova object from stats package in Pandoc's markdown.
}
