% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.nls}
\alias{pander.summary.nls}
\title{Prints an summary.nls object from stats package in Pandoc's markdown.}
\usage{
\method{pander}{summary.nls}(x, summary = TRUE,
  add.significance.stars = FALSE, digits = panderOptions("digits"),
  show.convergence = FALSE, ...)
}
\arguments{
\item{x}{an summary.nls object}

\item{summary}{(defaut:\code{TRUE}) if used for summary.lm or lm}

\item{add.significance.stars}{if significance stars should be shown for P value}

\item{digits}{number of digits of precision}

\item{show.convergence}{(defaut:\code{FALSE}) if to print convergence info}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an summary.nls object from stats package in Pandoc's markdown.
}
