% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.polr}
\alias{pander.summary.polr}
\title{Prints an summary.polr object from MASS package in Pandoc's markdown.}
\usage{
\method{pander}{summary.polr}(x, digits = panderOptions("digits"),
  round = panderOptions("round"),
  keep.trailing.zeros = panderOptions("keep.trailing.zeros"), ...)
}
\arguments{
\item{x}{an summary.polr object}

\item{digits}{number of digits of precision passed to format}

\item{round}{number of rounding digits passed to round}

\item{keep.trailing.zeros}{to show or remove trailing zeros in numbers on a column basis width}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an summary.polr object from MASS package in Pandoc's markdown.
}
