% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.summary.prcomp}
\alias{pander.summary.prcomp}
\title{Pander method for summary.prcomp class}
\usage{
\method{pander}{summary.prcomp}(x, caption = attr(x, "caption"),
  summary = TRUE, ...)
}
\arguments{
\item{x}{a summary.prcomp object}

\item{caption}{caption (string) to be shown under the table}

\item{summary}{(default:\code{TRUE}) if extended summary should be printed}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.prcomp object in Pandoc's markdown.
}

