% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.rms}
\alias{pander.summary.rms}
\title{Prints an summary.rms from rms package in Pandoc's markdown.}
\usage{
\method{pander}{summary.rms}(x, ...)
}
\arguments{
\item{x}{an summary.rms object}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an summary.rms from rms package in Pandoc's markdown.
}
