% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.survreg}
\alias{pander.summary.survreg}
\title{Prints an survreg object from survival package in Pandoc's markdown.}
\usage{
\method{pander}{summary.survreg}(x, summary = TRUE,
  digits = panderOptions("digits"), round = panderOptions("round"),
  keep.trailing.zeros = panderOptions("keep.trailing.zeros"), ...)
}
\arguments{
\item{x}{an survreg object}

\item{summary}{if summary should be printed}

\item{digits}{number of digits of precision passed to format}

\item{round}{number of rounding digits passed to round}

\item{keep.trailing.zeros}{to show or remove trailing zeros in numbers on a column basis width}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an survreg object from survival package in Pandoc's markdown.
}
