% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.summary.table}
\alias{pander.summary.table}
\title{Pander method for summary.table class}
\usage{
\method{pander}{summary.table}(x, caption = attr(x, "caption"),
  print.call = T, ...)
}
\arguments{
\item{x}{an function object}

\item{caption}{caption (string) to be shown under the table}

\item{print.call}{(defaut:\code{TRUE}) if call should be printed}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Renders an summary.table object in Pandoc's markdown.
}
