% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.survfit}
\alias{pander.survfit}
\title{Pander method for survfit class}
\usage{
\method{pander}{survfit}(x, caption = attr(x, "caption"), scale = 1,
  print.rmean = getOption("survfit.print.rmean"),
  rmean = getOption("survfit.rmean"), ...)
}
\arguments{
\item{x}{the result of a call to the survfit function.}

\item{caption}{caption (string) to be shown under the table}

\item{scale}{a numeric value to rescale the survival time, e.g., if the input data to survfit were in days, scale=365 would scale the printout to years.}

\item{print.rmean, rmean}{options for computation and display of the restricted mean}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an survfit object in Pandoc's markdown.
}
