% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.survreg}
\alias{pander.survreg}
\title{Prints an survreg object from survival package in Pandoc's markdown.}
\usage{
\method{pander}{survreg}(x, ...)
}
\arguments{
\item{x}{an survreg object}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an survreg object from survival package in Pandoc's markdown.
}
