% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.tabular}
\alias{pander.tabular}
\title{Pander method for tables::tabular class}
\usage{
\method{pander}{tabular}(x, caption = attr(x, "caption"),
  digits = panderOptions("digits"), ...)
}
\arguments{
\item{x}{a tables::tabular object}

\item{caption}{caption (string) to be shown under the table}

\item{digits}{number of digits of precision}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a tables::tabular object in Pandoc's markdown.
}

