% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.ts}
\alias{pander.ts}
\title{Pander method for timeseries class}
\usage{
\method{pander}{ts}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a timeseries object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a timeseries object in Pandoc's markdown.
}

