% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.vector}
\alias{pander.vector}
\title{Helper function to deal with atomic vectors}
\usage{
\method{pander}{vector}(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{ignored parameters}
}
\description{
Helper function to deal with atomic vectors
}
\keyword{internal}
