% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander_return}
\alias{pander_return}
\title{Pander and capture output}
\usage{
pander_return(...)
}
\arguments{
\item{...}{everything passed to \code{pander}}
}
\description{
This is a wrapper function around \code{pander} but instead of printing to \code{stdout}, this function returns a character vector of the captured lines.
}
\seealso{
pander
}
