% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.date.return}
\alias{pandoc.date.return}
\alias{pandoc.date}
\title{Dates}
\usage{
pandoc.date.return(x, inline = TRUE, simplified = FALSE, ...)
}
\arguments{
\item{x}{date or vector of dates}

\item{inline}{if to render vector of dates as inline paragraph or not (as list)}

\item{simplified}{if just add date formatting to vector of dates}

\item{...}{extra arguments passed by from parent call, disregarded}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Pandoc's mardown date.
}
\examples{
pandoc.date(Sys.Date())
pandoc.date(Sys.Date() - 1:10)
pandoc.date(Sys.Date() - 1:10, inline = FALSE)
}
