% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.formula.return}
\alias{pandoc.formula.return}
\alias{pandoc.formula}
\title{Formulas}
\usage{
pandoc.formula.return(x, text = NULL, max.width = 80, caption,
  add.line.breaks = FALSE, ...)
}
\arguments{
\item{x}{formula}

\item{text}{text to be written before result in the same line. Typically used by calls from other functions in the package}

\item{max.width}{maximum width in characters per line}

\item{caption}{caption (string) to be shown under the formula}

\item{add.line.breaks}{if to add 2 line breaks after formula}

\item{...}{extra arguments passed by from parent call, disregarded}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Pandoc's mardown formula.
}
\examples{
pandoc.formula(y ~ x)
pandoc.formula(formula(paste('y ~ ', paste0('x', 1:12, collapse = ' + '))))
}
