% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.horizontal.rule.return}
\alias{pandoc.horizontal.rule.return}
\alias{pandoc.horizontal.rule}
\title{Create horizontal rule}
\usage{
pandoc.horizontal.rule.return()
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Creates a Pandoc's markdown format horizontal line with trailing and leading newlines.
}
\references{
John MacFarlane (2012): _Pandoc User's Guide_. \url{http://johnmacfarlane.net/pandoc/README.html}
}
