% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.strong.return}
\alias{pandoc.strong.return}
\alias{pandoc.strong}
\title{Strong emphasis}
\usage{
pandoc.strong.return(x)
}
\arguments{
\item{x}{character vector}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Pandoc's markdown strong emphasis format (e.g. \code{**FOO**}) is added to character string.
}
\examples{
pandoc.strong('FOO')
pandoc.strong(c('FOO', '**FOO**'))
pandoc.strong.return('FOO')
}
\references{
John MacFarlane (2012): _Pandoc User's Guide_. \url{http://johnmacfarlane.net/pandoc/README.html}
}
\seealso{
\code{\link{pandoc.emphasis}} \code{\link{pandoc.strikeout}} \code{\link{pandoc.verbatim}}
}
