% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{path_to_pandoc}
\alias{path_to_pandoc}
\title{Find path to the pandoc binary by checking the \code{PATH} and the \code{RSTUDIO_PANDOC} env vars}
\usage{
path_to_pandoc()
}
\value{
file path
}
\description{
Find path to the pandoc binary by checking the \code{PATH} and the \code{RSTUDIO_PANDOC} env vars
}
