% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_color_pander}
\alias{scale_color_pander}
\alias{scale_colour_pander}
\alias{scale_fill_pander}
\title{Color scale from the pander package}
\usage{
scale_color_pander(...)

scale_colour_pander(...)

scale_fill_pander(...)
}
\arguments{
\item{...}{passed to \code{\link[ggplot2]{scale_colour_discrete}}}
}
\description{
The \pkg{pander} ships with a default colorblind and printer-friendly color
palette borrowed from \url{http://jfly.iam.u-tokyo.ac.jp/color/}.
}
\seealso{
\code{\link{theme_pander}}
}

