\name{set.caption}
\alias{set.caption}
\title{Adds caption in current block}
\usage{
  set.caption(x)
}
\arguments{
  \item{x}{string}
}
\value{
  A custom syntax _message_ like: \code{\\caption\{...\}}
}
\description{
  This is a helper function to be used inside brew blocks
  to add a caption to the returning image/table.
}

