% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_pander}
\alias{theme_pander}
\title{A ggplot theme for the pander package}
\usage{
theme_pander(base_size = panderOptions("graph.fontsize"),
  base_family = panderOptions("graph.fontfamily"),
  nomargin = panderOptions("graph.nomargin"),
  fc = panderOptions("graph.fontcolor"), gM = panderOptions("graph.grid"),
  gm = panderOptions("graph.grid.minor"),
  gc = panderOptions("graph.grid.color"),
  gl = panderOptions("graph.grid.lty"),
  boxes = panderOptions("graph.boxes"),
  bc = panderOptions("graph.background"),
  pc = panderOptions("graph.panel.background"),
  lp = panderOptions("graph.legend.position"),
  axis = panderOptions("graph.axis.angle"))
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{nomargin}{suppress the white space around the plot (boolean)}

\item{fc}{font color (name or hexa code)}

\item{gM}{major grid (boolean)}

\item{gm}{minor grid (boolean)}

\item{gc}{grid color (name or hexa code)}

\item{gl}{grid line type (\code{lty})}

\item{boxes}{to render a border around the plot or not}

\item{bc}{background color (name or hexa code)}

\item{pc}{panel background color (name or hexa code)}

\item{lp}{legend position}

\item{axis}{axis angle as defined in \code{par(les)}}
}
\description{
The \pkg{pander} ships with a default theme when the "unify plots" option is
enabled via \code{panderOptions}, which is now also available outside of \pkg{pander} internals,
like \code{evals}, \code{eval.msgs} or \code{Pandoc.brew}.
}
\examples{
library("ggplot2")
p <- qplot(mpg, wt, data = mtcars)
p + theme_pander()

panderOptions('graph.grid.color', 'red')
p + theme_pander()

p <- ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) + geom_point()
p + theme_pander() + scale_color_pander()

## standard examples of the ggtheme package
qplot(carat, price, data = diamonds, colour = cut) +
  theme_pander() +
  scale_colour_pander()

ggplot(diamonds, aes(clarity, fill = cut)) +
  geom_bar() +
  scale_fill_pander() +
  theme_pander()
}

