% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim.spaces}
\alias{trim.spaces}
\title{Trim leading and trailing spaces}
\usage{
trim.spaces(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Trim leading and trailing spaces
}
\seealso{
\code{trim.space} in \code{rapport} package
}
