% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add.lattice.xsubticks}
\alias{add.lattice.xsubticks}
\alias{add.lattice.ysubticks}
\title{Add subticks}
\usage{
add.lattice.xsubticks(lim, ..., n = 2)
}
\arguments{
\item{lim}{see \code{latticeExtra::xscale.components.subticks}}

\item{...}{see \code{latticeExtra::xscale.components.subticks}}

\item{n}{see \code{latticeExtra::xscale.components.subticks}}
}
\description{
See \code{latticeExtra::xscale.components.subticks}.
}
\references{
This is a simplified verwion of \code{latticeExtra::xscale.components.subticks}.
}
\keyword{internal}
