% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.alignment}
\alias{get.alignment}
\title{Get alignment}
\usage{
get.alignment(df)
}
\value{
vector of alignment parameters
}
\description{
Get alignment from temporary environment, truncating that and applying rownames and other columns alignment to passed \code{df}.
}
\keyword{internal}

