% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.alignment}
\alias{get.alignment}
\title{Get alignment}
\usage{
get.alignment(df, remove.obious.rownames = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{remove.obious.rownames}{if this rule should be applied or not}
}
\value{
vector of alignment parameters
}
\description{
Get alignment from temporary environment, truncating that and applying rownames and other columns alignment to passed \code{df}.
}
\keyword{internal}
