% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has.rownames}
\alias{has.rownames}
\title{Check if rownames are available}
\usage{
has.rownames(x)
}
\arguments{
\item{x}{a tabular-like R object}
}
\value{
\code{TRUE} OR \code{FALSE}
}
\description{
Dummy helper to check if the R object has real rownames or not.
}
