% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{openFileInOS}
\alias{openFileInOS}
\title{Open file}
\usage{
openFileInOS(f)
}
\arguments{
\item{f}{file (with full path)}
}
\description{
Tries to open a file with operating system's default program.
}
\references{
This function is a fork of David Hajage's \code{convert} function: \url{https://github.com/eusebe/ascii/blob/master/R/export.r}
}
