% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.anova}
\alias{pander.anova}
\title{Pander method for anova class}
\usage{
\method{pander}{anova}(x, caption = attr(x, "caption"),
  add.significance.stars = FALSE, ...)
}
\arguments{
\item{x}{an anova object}

\item{caption}{caption (string) to be shown under the table}

\item{add.significance.stars}{if significance stars should be shown for P value}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an anova object in Pandoc's markdown.
}
