% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.coxph}
\alias{pander.coxph}
\title{Pander method for coxph class}
\usage{
\method{pander}{coxph}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an coxph object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a coxph object in Pandoc's markdown.
}
