% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pander.CrossTable}
\alias{pander.CrossTable}
\title{Pander method for CrossTable class}
\usage{
\method{pander}{CrossTable}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a CrossTable object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a CrossTable object in Pandoc's markdown.
}

