% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.CrossTable}
\alias{pander.CrossTable}
\title{Pander method for CrossTable class}
\usage{
\method{pander}{CrossTable}(x, caption = attr(x, "caption"),
  digits = panderOptions("digits"), total.r = x$total.r,
  total.c = x$total.c, ...)
}
\arguments{
\item{x}{a CrossTable object}

\item{caption}{caption (string) to be shown under the table}

\item{digits}{number of digits of precision}

\item{total.r}{if to print row totals. Default values is taken from CrossTable object}

\item{total.c}{if to print column totals. Default values is taken from CrossTable object}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a CrossTable object in Pandoc's markdown.
}
