% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.Date}
\alias{pander.Date}
\title{Pander method for Date class}
\usage{
\method{pander}{Date}(x, ...)
}
\arguments{
\item{x}{a Date object}

\item{...}{optional parameters passed to raw \code{pandoc.date} function}
}
\description{
Prints a Date object in Pandoc's markdown.
}
