% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.describe}
\alias{pander.describe}
\title{Pander method for describe class}
\usage{
\method{pander}{describe}(x, caption = attr(x, "caption"),
  digits = panderOptions("digits"), ...)
}
\arguments{
\item{x}{an describe object}

\item{caption}{caption (string) to be shown under the table}

\item{digits}{number of digits of precision}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a describe object in Pandoc's markdown.
}
