% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.evals}
\alias{pander.evals}
\title{Pander method for evals class}
\usage{
\method{pander}{evals}(x, ...)
}
\arguments{
\item{x}{a evals object}

\item{...}{ignored parameters}
}
\description{
Prints a evals object in Pandoc's markdown.
}
