% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.image}
\alias{pander.image}
\title{Pander method for image class}
\usage{
\method{pander}{image}(x, caption = attr(x, "caption"), href = attr(x,
  "href"), ...)
}
\arguments{
\item{x}{a image object}

\item{caption}{caption (string) to be shown under the table}

\item{href}{link that image should be linked with}

\item{...}{ignored parameters}
}
\description{
Prints a image object in Pandoc's markdown.
}
