% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.lm}
\alias{pander.lm}
\title{Pander method for summary.lm class}
\usage{
\method{pander}{lm}(x, caption = attr(x, "caption"), covariate.labels,
  omit, ...)
}
\arguments{
\item{x}{a summary.glm object}

\item{caption}{caption (string) to be shown under the table}

\item{covariate.labels}{vector to replace covariate lables in the table}

\item{omit}{vector of variable to omit for priting in resulting table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a summary.lm object in Pandoc's markdown.
}
