% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.lme}
\alias{pander.lme}
\title{Pander method for lme class}
\usage{
\method{pander}{lme}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a lme object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a lme object in Pandoc's markdown.
}
