% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.NULL}
\alias{pander.NULL}
\title{Pander method for a NULL object}
\usage{
\method{pander}{NULL}(x, ...)
}
\arguments{
\item{x}{a NULL object}

\item{...}{ignored parameters}
}
\description{
Prints a NULL object in Pandoc's markdown.
}
