% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.numeric}
\alias{pander.numeric}
\title{Pander method for numeric class}
\usage{
\method{pander}{numeric}(x, ...)
}
\arguments{
\item{x}{a numeric object}

\item{...}{igroned parameter}
}
\description{
Prints a numeric class in Pandoc's markdown.
}
