% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{pander.option}
\alias{pander.option}
\title{Deprecated panderOptions}
\usage{
pander.option(x, ...)
}
\arguments{
\item{x}{passed to \code{panderOptions}}

\item{...}{passed to \code{panderOptions}}
}
\description{
Deprecated panderOptions
}
\seealso{
panderOptions
}

