% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.POSIXct}
\alias{pander.POSIXct}
\title{Pander method for POSIXct class}
\usage{
\method{pander}{POSIXct}(x, ...)
}
\arguments{
\item{x}{a POSIXct object}

\item{...}{optional parameters passed to raw \code{pandoc.date} function}
}
\description{
Prints a POSIXct object in Pandoc's markdown.
}
