% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.rapport}
\alias{pander.rapport}
\title{Pander method for rapport class}
\usage{
\method{pander}{rapport}(x, ...)
}
\arguments{
\item{x}{a rapport object}

\item{...}{ignored parameters}
}
\description{
Prints a rapport object in Pandoc's markdown.
}
