% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.rlm}
\alias{pander.rlm}
\title{Pander method for rlm class}
\usage{
\method{pander}{rlm}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an rlm object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an rlm object in Pandoc's markdown.
}
