% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.smooth.spline}
\alias{pander.smooth.spline}
\title{Pander method for smooth.spline class}
\usage{
\method{pander}{smooth.spline}(x, ...)
}
\arguments{
\item{x}{an smooth.spline object}

\item{...}{igroned parameters}
}
\description{
Prints an smooth.spline object in Pandoc's markdown.
}
