% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.survdiff}
\alias{pander.survdiff}
\title{Pander method for survdiff class}
\usage{
\method{pander}{survdiff}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an survdiff object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints an survdiff object in Pandoc's markdown.
}
