% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.tabular}
\alias{pander.tabular}
\title{Pander method for tabular class}
\usage{
\method{pander}{tabular}(x, ...)
}
\arguments{
\item{x}{an tabular object}

\item{...}{ignored parameters}
}
\description{
Pander method for tabular class
}

