% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.tabular}
\alias{pander.tabular}
\title{Pander method for tabular class}
\usage{
\method{pander}{tabular}(x, caption = attr(x, "caption"),
  emphasize.rownames = TRUE, digits = panderOptions("digits"), ...)
}
\arguments{
\item{x}{an function object}

\item{caption}{caption (string) to be shown under the table}

\item{emphasize.rownames}{(defaut:\code{TRUE}) if rownames should be highlighted}

\item{digits}{number of digits of precision}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Renders an tabular object in Pandoc's markdown.
}
