% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pander.zoo}
\alias{pander.zoo}
\title{Pander method for zoo class}
\usage{
\method{pander}{zoo}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an zoo object}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw \code{pandoc.table} function}
}
\description{
Prints a zoo object in Pandoc's markdown.
}
