% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.image.return}
\alias{pandoc.image.return}
\alias{pandoc.image}
\title{Create pandoc image tags}
\usage{
pandoc.image.return(img, caption = storage$caption)
}
\arguments{
\item{img}{image path}

\item{caption}{text}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
Creates a Pandoc's markdown format image hyperlink.
}
\note{
The \code{caption} text is read from an internal buffer which defaults to \code{NULL}. To update that, call \code{link{set.caption}} before.
}
\examples{
pandoc.image('foo.png')
pandoc.image('foo.png', 'Nice image, huh?')
}
\references{
John MacFarlane (2012): _Pandoc User's Guide_. \url{http://johnmacfarlane.net/pandoc/README.html}
}
\seealso{
\code{\link{set.caption}}
}
