% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals.R
\name{redraw.recordedplot}
\alias{redraw.recordedplot}
\title{Redraws plot saved in file}
\usage{
redraw.recordedplot(file)
}
\arguments{
\item{file}{path and name of an rds file containing a plot object to be redrawn}
}
\description{
This function is a wrapper around \code{redrawPlot}.
}
\references{
Thanks to Jeroen Ooms \url{https://stat.ethz.ch/pipermail/r-devel/2012-January/062973.html}, JJ Allaire \url{https://github.com/rstudio/rstudio/commit/eb5f6f1db4717132c2ff111f068ffa6e8b2a5f0b}, and Gabriel Becker.
}
\seealso{
\code{\link{evals}}
}
