% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals.R
\name{redrawPlot}
\alias{redrawPlot}
\title{Redraw a recordedplot, grid, trellis, or ggplot2 plot.}
\usage{
redrawPlot(rec_plot)
}
\arguments{
\item{rec_plot}{the plot object to redraw}
}
\description{
This function redraws the plot represented by \code{rec_plot}. It can redraw grid/trellis/ggplot2/etc plots, as well as \code{recordedplot} objects. For \code{recordedplot} objects it acts as a wrapper around \code{replayPlot} with memory tweaks to fix native symbol address errors when the recordedplot was loaded from an rda/rds file.
}
\references{
Thanks to Jeroen Ooms \url{https://stat.ethz.ch/pipermail/r-devel/2012-January/062973.html}, JJ Allaire \url{https://github.com/rstudio/rstudio/commit/eb5f6f1db4717132c2ff111f068ffa6e8b2a5f0b}, and Gabriel Becker.
}
\seealso{
\code{\link{redraw.recordedplot}}
}
