% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove.extra.newlines}
\alias{remove.extra.newlines}
\title{Remove more then two joined newlines}
\usage{
remove.extra.newlines(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Remove more then two joined newlines
}
\examples{
remove.extra.newlines(c('\\n\\n\\n', '\\n\\n', '\\n'))
}
