% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repChar}
\alias{repChar}
\title{Repeating chars}
\usage{
repChar(x, n, sep = "")
}
\arguments{
\item{x}{string to repeat}

\item{n}{integer}

\item{sep}{separator between repetitions}
}
\value{
character vector
}
\description{
Repeating a string \code{n} times and returning a concatenated character vector.
}

