% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set.alignment}
\alias{set.alignment}
\title{Sets alignment for tables}
\usage{
set.alignment(default = panderOptions("table.alignment.default"),
  row.names = panderOptions("table.alignment.rownames"),
  permanent = FALSE)
}
\arguments{
\item{default}{character vector which length equals to one (would be repeated \code{n} times) ot \code{n} - where \code{n} equals to the number of columns in the following table}

\item{row.names}{string holding the alignment of the (optional) row names}

\item{permanent}{(default \code{FALSE}) if alignment is permanent (for all future tables) or not. It's cleaner to use \code{panderOptions} instead.}
}
\description{
This is a helper function to update the alignment (\code{justify} parameter in \code{pandoc.table}) of the next returning table. Possible values are: \code{centre} or \code{center}, \code{right}, \code{left}.
}
