% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set.caption}
\alias{set.caption}
\title{Adds caption in current block}
\usage{
set.caption(x, permanent = FALSE)
}
\arguments{
\item{x}{string}

\item{permanent}{(default \code{FALSE}) if caption is permanent (for all future tables) or not}
}
\description{
This is a helper function to add a caption to the returning image/table.
}
